done task_23

+++ service/src/main/java/com/itclopedia/courses/services/CategoryService.java
@@ -1,7 +1,9 @@
 package com.itclopedia.courses.services;
 
 import com.itclopedia.courses.dao.CategoryRepository;
+import com.itclopedia.courses.exceptions.EntityAlreadyExistsException;
 import com.itclopedia.courses.models.Category;
+import com.itclopedia.courses.exceptions.EntityNotFoundException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -18,19 +20,27 @@ public class CategoryService {
     }
 
     public void addCategory(Category category) {
+        if (categoryRepository.existsById(category.getId())) {
+            throw new EntityAlreadyExistsException("Category", category.getId());
+        }
         categoryRepository.save(category);
     }
-
     public void updateCategory(Category category) {
+        if (!categoryRepository.existsById(category.getId())) {
+            throw new EntityNotFoundException("Category", category.getId());
+        }
         categoryRepository.save(category);
     }
 
     public void deleteCategory(int categoryId) {
+        if (!categoryRepository.existsById(categoryId)) {
+            throw new EntityNotFoundException("Category", categoryId);
+        }
         categoryRepository.deleteById(categoryId);
-
     }
 
     public Category getCategoryById(int categoryId) {
-        return categoryRepository.findById(categoryId).orElse(null);
+        return categoryRepository.findById(categoryId)
+                .orElseThrow(() -> new EntityNotFoundException("Category", categoryId));
     }
 }
